Fix a bug where fw or is would prove correctness of unsafe programs
The steps would send the step to confirm it with the base case, which
would be set as the max number of loops to iterate.

But, depending on the increment of the loop, the next iteration could
increment the step past the max number of iterations previously set.

This was the lazy solution: set the max number of iterations as
being the number sent by the other steps + the increment.

Signed-off-by: Mikhail Gadelha <mikhail.ramalho@gmail.com>